fixed Chart Logic

+++ BikeStat/BikeStat/PresentationLayer/Modules/Statistics/StatisticsChartDataModel.swift
@@ -8,7 +8,7 @@ struct StatisticsChartDataModel: Identifiable, Equatable {
     let id = UUID()
     var number: Int
     var title: String
-    var distance: Int
+    var distance: Double
     var complexity: String?
 }
